# public/docs/openapi.yaml
openapi: 3.0.3
info:
    title: Laravel
    description: ''
    version: 1.0.0
servers:
    -
        url: 'http://localhost'
paths:
...
    '/orders/{order}':
        get:
            summary: 'Display the specified order.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200'
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: "{\n\"order_id\":35000,\n\"invoice_id\":2018023,\n\"domain\":'www.stylka.cz',\n\"firstname\":'Jan',\n\"lastname\":'Zelený',\n\"comment\":'Objednávka zaplacena.',\n\"order_status\":'Odesláno dopravcem'\n\"shipping_method\":'Zásilkovna',\n\"label\":1,\n\"date_added\":2018-07-13,\n\"total\":453,\n\"payment_status\":1,\n\"payment_status\" => 1,\n\"profit\" => 420,\n\"slovakia\" => 0,\n\"instock\" => 0,\n\"referrer\" => 'Google',\n\"agree_gdpr\" => 1,\n\"payment_method\" => 'Platba kartou',\n\"email\" => 'o@gmail.com',\n\"telephone\" => '+420555111444',\n}"
            tags:
                - Order
            security: []
        put:
            summary: 'Update the specified order in storage.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    domain: true
                                    currency: true
                                    noTaxTotal: 100
                                    tax: 21
                                    total: 121
                                properties:
                                    domain: { type: boolean, example: true }
                                    currency: { type: boolean, example: true }
                                    noTaxTotal: { type: integer, example: 100 }
                                    tax: { type: integer, example: 21 }
                                    total: { type: integer, example: 121 }
            tags:
                - Order
            requestBody:
                required: false
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                firstname:
                                    type: string
                                    description: ''
                                    example: architecto
                                lastname:
                                    type: string
                                    description: ''
                                    example: et
                                email:
                                    type: string
                                    description: ''
                                    example: et
                                telephone:
                                    type: string
                                    description: ''
                                    example: et
                                shipping_firstname:
                                    type: string
                                    description: ''
                                    example: error
                                shipping_lastname:
                                    type: string
                                    description: ''
                                    example: omnis
                                shipping_address_1:
                                    type: string
                                    description: ''
                                    example: illum
                                shipping_address_2:
                                    type: string
                                    description: ''
                                    example: dicta
                                shipping_city:
                                    type: string
                                    description: ''
                                    example: eos
                                shipping_postcode:
                                    type: string
                                    description: ''
                                    example: eos
                                shipping_country:
                                    type: string
                                    description: ''
                                    example: velit
                                payment_firstname:
                                    type: string
                                    description: ''
                                    example: quam
                                payment_lastname:
                                    type: string
                                    description: ''
                                    example: aperiam
                                payment_address_1:
                                    type: string
                                    description: ''
                                    example: eveniet
                                payment_address_2:
                                    type: string
                                    description: ''
                                    example: aperiam
                                payment_city:
                                    type: string
                                    description: ''
                                    example: error
                                payment_postcode:
                                    type: string
                                    description: ''
                                    example: vel
                                payment_country:
                                    type: string
                                    description: ''
                                    example: rem
                                domain:
                                    type: string
                                    description: ''
                                    example: '"www.stylka.cz"'
                                currency:
                                    type: string
                                    description: ''
                                    example: ratione
                                customer_id:
                                    type: integer
                                    description: ''
                                    example: 15
                                language:
                                    type: string
                                    description: ''
                                    example: quibusdam
                                company:
                                    type: string
                                    description: ''
                                    example: sunt
                                comment_c:
                                    type: string
                                    description: 'The comment written by the customer.'
                                    example: aut
                                comment_e:
                                    type: string
                                    description: 'The comment written by somebody else than customer.'
                                    example: minus
                                notify:
                                    type: string
                                    description: ''
                                    example: nisi
                                order_status:
                                    type: string
                                    description: ''
                                    example: '"Nevyřízeno."'
                                shipping_method:
                                    type: string
                                    description: ''
                                    example: '"Zásilkovna"'
                                payment_status:
                                    type: integer
                                    description: 'Whether the order was paid.'
                                    example: 1
                                referrer:
                                    type: string
                                    description: 'The source website where the customer has come from.'
                                    example: officia
                                agree_gdpr:
                                    type: integer
                                    description: 'Whether the customer agrees with gdpr policy.'
                                    example: 1
                                payment_method:
                                    type: string
                                    description: ''
                                    example: provident
                                fax:
                                    type: string
                                    description: ''
                                    example: facere
                                regNum:
                                    type: string
                                    description: 'In Czech language IČO.'
                                    example: quia
                                taxRegNum:
                                    type: string
                                    description: 'In Czech language DIČO.'
                                    example: facere
                                ip:
                                    type: string
                                    description: ''
                                    example: reprehenderit
                                reason:
                                    type: string
                                    description: ''
                                    example: '"Reklamace"'
                                wrong_order_id:
                                    type: integer
                                    description: ''
                                    example: 6
                                competition:
                                    type: integer
                                    description: ''
                                    example: 7
                                euVAT:
                                    type: integer
                                    description: ''
                                    example: 14
                                viewed:
                                    type: integer
                                    description: 'Whether the order has been viewed.'
                                    example: 0
                                shipping_company:
                                    type: string
                                    description: ''
                                    example: itaque
                                shipping_zone:
                                    type: string
                                    description: ''
                                    example: possimus
                                shipping_address_format:
                                    type: string
                                    description: ''
                                    example: harum
                                payment_company:
                                    type: string
                                    description: ''
                                    example: nulla
                                payment_zone:
                                    type: string
                                    description: ''
                                    example: omnis
                                payment_address_format:
                                    type: string
                                    description: ''
                                    example: sit
            security: []
        delete:
            summary: 'Remove the specified order from storage.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200'
                    content:
                        application/json:
                            schema:
                                type: boolean
                                example: true
            tags:
                - Order
            security: []
        parameters:
            -
                in: path
                name: order
                description: 'order id'
                example: '35022'
                required: true
                schema:
                    type: string
...
    '/orders/{order}/products/{product}':
        get:
            summary: 'Display the specified ordered product.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200'
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: "{\n\"order_product_id:112,\n\"order_id\":50,\n\"product_id\":183,\n\"name\":\"Náhradní hřebeny SP-HC5000\",\n\"model\":\"4008496717552\",\n\"price\":207.5000,\n\"total\":207.5000,\n\"tax\":21.0000,\n\"quantity\":1,\n\"sort_order\":0,\n\"is_transfer\":0,\n\"is_action\":0,\n\"purchase_price\":131.8200,\n\"warranty\":24,\n\"gift\":0\n}"
            tags:
                - Order_product
            security: []
        put:
            summary: 'Update the specified ordered product in storage.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    quantity: false
                                properties:
                                    quantity: { type: boolean, example: false }
            tags:
                - Order_product
            requestBody:
                required: false
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                quantity:
                                    type: integer
                                    description: ''
                                    example: 5
                                sort_order:
                                    type: integer
                                    description: ''
                                    example: 1
                                is_action:
                                    type: integer
                                    description: ''
                                    example: 1
            security: []
        delete:
            summary: 'Remove the specified ordered product from storage.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    noTaxTotal: 100
                                    tax: 21
                                    total: 121
                                properties:
                                    noTaxTotal: { type: integer, example: 100 }
                                    tax: { type: integer, example: 21 }
                                    total: { type: integer, example: 121 }
            tags:
                - Order_product
            security: []
        parameters:
            -
                in: path
                name: order
                description: 'order id'
                example: '35022'
                required: true
                schema:
                    type: string
            -
                in: path
                name: product
                description: ''
                example: itaque
                required: true
                schema:
                    type: string
            -
                in: path
                name: order_product
                description: 'order product id'
                example: '74850'
                required: true
                schema:
                    type: string
...